Split smer lookups

+++ LOGAN-Graph-Native/src/graph/smer.h
@@ -87,12 +87,12 @@ int smerEntryCompar(const void *ptr1, const void *ptr2);
 //SmerId shuffleSmerIdLeft(SmerId id, u8 base);
 
 void calculatePossibleSmers(u8 *data, s32 maxIndex, SmerId *smerIds);
-void calculatePossibleSmersComp(u8 *data, s32 maxIndex, SmerId *smerIds, u32 *compFlags);
+void calculatePossibleSmersComp(u8 *data, s32 maxIndex, SmerId *smerIds, u8 *compFlags);
 //void calculatePossibleSmers2(u8 *data, s32 maxIndex, SmerId *smerIds, u32 *compFlags);
 
 u64 hashForSmer(SmerEntry entry);
 u32 sliceForSmer(SmerId smer, u64 hash);
-SmerId recoverSmerId(u64 sliceNum, SmerEntry entry);
+SmerId recoverSmerId(u32 sliceNum, SmerEntry entry);
 
 
 /********************** SmerMap functionality **********************/
